Refactor so DesktopGL tests, don't break other platforms.
