+++ Tests/MonoGame.Tests.DesktopVK.csproj
@@ -27,6 +27,7 @@
     <Compile Remove="Runner/iOS/**/*" />
 
     <Compile Remove="Framework/SDL/**/*" />
+    <Compile Remove="Framework/Devices/Power/PowerStatusTest.DesktopGL.cs" />
   </ItemGroup>
 
   <ItemGroup>
